requirements.txt
# Core LLM clients
openai>=1.50.0             # OpenRouter uses OpenAI-compatible API
anthropic>=0.40.0          # optional: direct Anthropic fallback

# LlamaIndex core + integrations
llama-index-core>=0.12.0
llama-index-vector-stores-chroma>=0.5.0    # local vector store (default)
llama-index-vector-stores-supabase>=0.3.0  # cloud vector store (optional)
llama-index-embeddings-openai>=0.3.0       # optional: paid embeddings
llama-index-embeddings-huggingface>=0.5.0  # free local embeddings (default)
llama-index-llms-anthropic>=0.10.0
llama-index-readers-json>=0.4.0
llama-index-readers-file>=0.4.0

# Vector stores
chromadb>=1.5.0            # local persistent vector DB
vecs>=0.4.0                # Supabase pgvector client (optional)

# Utilities
python-dotenv>=1.0.0
